% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animal_observations.R
\docType{data}
\name{animal_observations}
\alias{animal_observations}
\title{Animal observation records from point surveys}
\format{
Dataframe
}
\source{
TBC
}
\usage{
data(animal_observations)
}
\description{
Time-based point observations of animals (years 2016 to 2022)
at sampling point locations within six towns in Singapore:
Punggol (PG), Queenstown (QT), Tampines (TP), Jurong West (JW),
Bishan (BS) and Woodlands (WL). Species abundance and occurrence
were recorded for four types of animal surveys, corresponding to four
animal groups (taxa): Birds, butterflies, odonates and amphibians.
}
\examples{
head(animal_observations)

}
\references{
TBC
}
\seealso{
\itemize{
\item \link{animal_surveys} containing information about the surveys conducted.
\item \link{sampling_points} where these animal surveys were conducted.
\item \link{sampling_areas} where the sampling points were generated within.
}
}
\keyword{datasets}
