% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-area.R
\docType{data}
\name{area}
\alias{area}
\title{Survey area boundaries}
\format{
An object of class 'list'.
}
\source{
TBC
}
\usage{
data(area)
}
\description{
Desc
}
\examples{
data(area)


}
\references{
TBC
}
\keyword{datasets}
