% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booter.R
\name{booter}
\alias{booter}
\title{Helper function to generate bootstrapped statistics of accuracy metrics}
\usage{
booter(df_predictions, n)
}
\arguments{
\item{df_predictions}{Dataframe of predictions from the function \code{validate_newdata()}
(output of \code{MuMIn::get.models()}).}

\item{n}{Number of repetitions}
}
\description{
Helper function to generate bootstrapped statistics of accuracy metrics
}
