% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_manual.R
\name{calc_manual}
\alias{calc_manual}
\title{Calculate metrics based on manually generated data at point locations}
\usage{
calc_manual(
  vector,
  name = NULL,
  points,
  buffer_sizes,
  plant_species = "species",
  building_levels = "levels",
  road_lanes = "lanes"
)
}
\arguments{
\item{vector}{sf dataframe containing geometric features (points, polygons or lines) to be summarised.
Coordinate reference system should be similar to \code{points}.}

\item{name}{Specify either \code{"buildings"}, \code{"roads"}, \code{"trees"}, \verb{"shrubs'}, \code{"turf"}, \code{"natveg"}, or \code{"water"}.
Used to represent the type of data in \code{vector}, each to be processed differently.}

\item{points}{Point locations (sf object) to calculate the metrics.}

\item{buffer_sizes}{Radius of circle (in mapunits) for each point location;
metrics will be calculated within the buffer area.}

\item{plant_species}{Column name in \code{vector} for plant species names,
if input \code{vector} is vegetation (i.e., argument \verb{name=} \code{"trees"} or \code{"shrubs"}).
Defaults to \code{"species"}. Column data should be of type character.}

\item{building_levels}{Column name in \code{vector} for the number of building levels,
when \code{name="buildings"}. Defaults to \code{"levels"}. Column data should be numeric.}

\item{road_lanes}{Column name in \code{vector} for the number of lanes per road line,
when \code{name="roads"}. Defaults to \code{"lanes"}. Column data should be numeric.}
}
\value{
A list containing the features/metrics calculated for \code{points}, appended as new columns.
Each element in the list corresponds to a particular buffer size.
}
\description{
Calculate metrics from manually generated vector data,
at specific point locations and buffer radii.
Currently supports vector data of buildings (polygons), roads (lines),
trees (points), shrubs (polygons), turf (polygons), natural vegetation (polygons),
and water (polygons).
}
\details{
Summaries will be calculated for each of the supplied layers (\verb{layer_<layername>})
and appended to the \code{points} data as additional columns.
}
