% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_osm.R
\name{calc_osm}
\alias{calc_osm}
\title{Calculate metrics based on OpenStreetMap data at point locations}
\usage{
calc_osm(
  vector,
  name = NULL,
  points,
  buffer_sizes,
  building_ndsm = NULL,
  building_height = "height",
  building_levels = "levels",
  road_lanes = "lanes"
)
}
\arguments{
\item{vector}{sf dataframe of either buildings (polygons) or roads (lines).}

\item{name}{Specify either \code{"buildings"} or \code{"roads"}.}

\item{points}{Point locations (sf object) to calculate the metrics.}

\item{buffer_sizes}{Radius of circle (in mapunits) for each point location;
metrics will be calculated within the buffer area.}

\item{building_ndsm}{SpatRaster object (\code{terra::rast()}) (optional). A continuous raster
of the normalised Digital Surface Model, used to calculate building heights.
If absent (\code{NULL}) and the variable is named in \code{predictors_osm}, the column \code{building_height} is used instead.
Defaults to \code{NULL}.}

\item{building_height}{Column name in \code{vector} for building height.
Defaults to \code{"height"}.}

\item{building_levels}{Column name in \code{vector} for the number of building levels.
Defaults to \code{"levels"}. Column data should be numeric.}

\item{road_lanes}{Column name in \code{vector} for the number of lanes per road line.
Defaults to \code{"lanes"}. Column data should be numeric.}
}
\value{
A list containing the features/metrics calculated for \code{points},  appended as new columns.
Each element in the list corresponds to a particular buffer size.
}
\description{
Calculate metrics from OpenStreetMap (OSM) vector data,
at specific point locations and buffer radii.
Currently supports vector data of buildings (polygons) and roads (lines).
}
