% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_specific_osm.R
\name{calc_specific_osm}
\alias{calc_specific_osm}
\title{Calculate specific metrics from OpenStreetMap data at point locations}
\usage{
calc_specific_osm(
  vector,
  building_ndsm = NULL,
  building_height = "height",
  building_levels = "levels",
  road_lanes = "lanes",
  predictors_osm,
  points,
  point_id = "point_id"
)
}
\arguments{
\item{vector}{sf dataframe of either buildings (polygons) or roads (lines).}

\item{building_ndsm}{SpatRaster object (\code{terra::rast()}) (optional). A continuous raster
of the normalised Digital Surface Model, used to calculate building heights.
If absent (\code{NULL}) and the variable is named in \code{predictors_osm}, the column \code{building_height} is used instead.
Defaults to \code{NULL}.}

\item{building_height}{Column name in \code{vector} for building height.
Defaults to \code{"height"}.}

\item{building_levels}{Column name in \code{vector} for the number of building levels.
Defaults to \code{"levels"}.}

\item{road_lanes}{Column name in \code{vector} for the number of lanes per road line.}

\item{predictors_osm}{Vector (character) of predictor variables to be calculated from the vector file(s).
The naming format is \verb{<radius in metres>_osm_<metric>} (e.g. \code{r50m_osm_buildingFA_ratio}).}

\item{points}{Points locations (sf object) to calculate the metrics.}

\item{point_id}{Column name of the sampling point id within the \code{points} sf. Defaults to \code{"point_id"}.}
}
\value{
The \code{points} object including new columns for the variables specified in \code{predictors_osm}.
}
\description{
Calculate specific metrics from OpenStreetMap (OSM) vector data at point locations.
The character vector of predictor names include the specified buffer radii within which to summarise each metric.
Currently supports vector data of buildings (polygons) and roads (lines).
}
