% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_taxongrps.R
\name{check_taxongrps}
\alias{check_taxongrps}
\title{Check taxon group-level records for subsequent removal}
\usage{
check_taxongrps(
  observations,
  level,
  species = "species",
  genus = "genus",
  family = "family",
  town = "town",
  round = "round",
  point_id = "point_id"
)
}
\arguments{
\item{observations}{Dataframe of species observations.
It should include columns for \code{species}, \code{genus}, \code{family}, \code{town}, \code{round}, and \code{point_id}.}

\item{level}{Specify whether to tally by \code{town} or \code{point}.}

\item{species}{Column name of the species name recorded in \code{observations}.
May include group-level records (\code{genus} or \code{family}).}

\item{genus}{Column name of the genus name for recorded \code{observations}.}

\item{family}{Column name of the family name for recorded \code{observations}.}

\item{town}{Column name of the town specified in \code{observations}.
Defaults to \code{town}.}

\item{round}{Column name of the sampling round specified in \code{observations}.
Defaults to \code{round}.}

\item{point_id}{Column name of the unique identifier for each point in \code{observations}. Defaults to \code{point_id}.}
}
\value{
Dataframe containing the taxon group names where all species within the group are observed.
Includes columns for the \code{round} and \code{town}/\code{point_id} (depending on argument \code{level}), as well as
the number (\code{n}) of species in the particular taxon group.
}
\description{
Tally and list all taxon groups where all species within the group are observed,
either at the level of individual towns or points (and survey round).
}
