% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_image_binary.R
\name{classify_image_binary}
\alias{classify_image_binary}
\title{Binary classification of a single-band image based on a threshold value}
\usage{
classify_image_binary(
  image,
  threshold = "otsu",
  range = NULL,
  levels = 256,
  file = NULL,
  ...
)
}
\arguments{
\item{image}{Single band raster to be classified (\code{SpatRaster} object from \code{library(terra)}),
or a file path to the image to be imported (\code{terra::rast()} will be used).}

\item{threshold}{Either a numeric value for a manually-specified threshold
indicating absence/presence, or \code{"otsu"} to calculate the threshold via Otsu's method.
Defaults to \code{"otsu"}.}

\item{range}{Numeric vector (of length 2) specifying the histogram range to be used
for Otsu's thresholding (only relevant if \code{threshold = "otsu"}). Defaults to the minimum and maximum values of the imported raster.}

\item{levels}{Number of histogram bins used to calculate the threshold value (only relevant if \code{threshold = "otsu"}).
Typically based on the bit depth of the image (e.g. 8-bit image has \code{2^8} = \code{256} levels).
Defaults to \code{256}.}

\item{file}{File path to export output raster. Defaults to \code{NULL}.}

\item{...}{Other arguments that may be supplied to \code{terra::writeRaster()}.}
}
\value{
The classified raster (\code{SpatRaster} object).
}
\description{
Classify a single-band image such as a spectral index or greyscale raster derived from satellite data.
If a threshold value is not specified, the Otsu's threshold can be calculated and used for classification.
}
\examples{
\dontrun{
  ndvi_mosaic <- system.file("extdata", "ndvi_mosaic.tif", package="biodivercity")
  terra::plot(ndvi_mosaic) # examine continuous raster
  ndvi_classified <- classify_image_binary(image = ndvi_mosaic)
  terra::plot(ndvi) # examine classified raster
}

}
\seealso{
\itemize{
\item \code{\link[=threshold_otsu]{threshold_otsu()}} Internal function used to calculate threshold value via Otsu's method.
}
}
