% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_mosaic.R
\name{classify_mosaic}
\alias{classify_mosaic}
\title{Binary classification of a spectral index mosaic image with a threshold value.}
\usage{
classify_mosaic(x, threshold_value = NULL, otsu = FALSE, output)
}
\arguments{
\item{x}{Spectral index mosaic image (.tif) to be classified.}

\item{threshold_value}{Numeric value to determine absence or presence of elements according to spectral index.}

\item{otsu}{TRUE or FALSE to perform Otsu's thresholding.}

\item{output}{Output filename.}
}
\value{
The .tif raster \verb{classified_<index>.tif} in the output folder.
}
\description{
Binary classification of a spectral index mosaic image with a threshold value.
If a threshold value is not specified, the Otsu's threshold can be calculated and used for classification.
}
