% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_simulator.R
\name{exclude_simulator}
\alias{exclude_simulator}
\title{Simulate multiple scenarios with animal surveys randomly excluded}
\usage{
exclude_simulator(
  community_matrix,
  survey_ref,
  exclude_level = "survey",
  exclude_num,
  specify_areas = NULL,
  specify_points = NULL,
  specify_periods = NULL,
  specify_cycles = NULL,
  area_ignore = NULL,
  period_ignore = NULL,
  cycle_ignore = NULL,
  survey_id = "survey_id",
  point_id = "point_id",
  area = "area",
  period = "period",
  cycle = "cycle",
  rep = 100
)
}
\arguments{
\item{community_matrix}{Dataframe of the species community matrix (one survey per row).
It should a column containing the unique IDs for the animal surveys conducted (e.g. \code{survey_id}).}

\item{survey_ref}{Dataframe containing information of the surveys conducted
(e.g. \code{survey_id}, \code{point_id}, \code{period}, \code{area}, \code{cycle}). Values in the column
\code{survey_id} should correspond to those in \code{community_matrix}.}

\item{exclude_level}{Whether to exclude based on \code{"survey"} or \code{"point"}. Defaults to \code{"survey"}.}

\item{exclude_num}{Number of surveys/points to exclude (remove) from the full dataset \code{community_matrix}.
Value should be a single integer.}

\item{specify_areas}{Specify the \code{area}(s) to randomly exclude surveys from (character string or vector).
Defaults to \code{NULL}, which excludes surveys randomly across all areas.}

\item{specify_points}{Specify the \code{point_id}(s) of sampling points to randomly exclude surveys from (character string or vector).
Defaults to \code{NULL}, which excludes surveys randomly across all points.}

\item{specify_periods}{Specify the survey \code{period}(s) to exclude surveys from (character string or vector). Defaults to \code{NULL},
which excludes surveys randomly across all periods.}

\item{specify_cycles}{Specify the survey \code{cycle}(s) to exclude surveys from (character string or vector). Defaults to \code{NULL},
which excludes surveys randomly across all cycles.}

\item{area_ignore}{Ignore specified \code{area}(s) from the simulation and output (character string or vector). Defaults to \code{NULL},
which includes all areas in the simulation.}

\item{period_ignore}{Ignore specified \code{period} from the simulation and output (character string or vector). Defaults to \code{NULL},
which includes all periods in the simulation.}

\item{cycle_ignore}{Ignore specified \code{cycles}(s) from the simulation and output (character string or vector). Defaults to \code{NULL},
which includes all cycles in the simulation.}

\item{survey_id}{Column name of the unique identifier for each survey in \code{community_matrix} and
\code{survey_ref}. Defaults to \code{survey_id}.}

\item{point_id}{Column name of the unique identifier for each sampling point in \code{survey_ref}.
Defaults to \code{survey_id}.}

\item{area}{Column name of the area specified in \code{survey_ref}.
Defaults to \code{area}.}

\item{period}{Column name of the survey period specified in \code{survey_ref}.
Defaults to \code{period}.}

\item{cycle}{Column name of the survey cycle specified in \code{survey_ref}.
Defaults to \code{cycle}.}

\item{rep}{Specify number of repetitions (scenarios) to simulate. Defaults to \code{100}.}
}
\value{
A dataframe containing the results of each iteration (column).
The count of sampling points with each species (row) is shown. The column \code{full}
shows the breakdown of counts within the original dataset.
}
\description{
Randomly exclude either animal surveys or sampling points (and all their associated surveys) from a
species community matrix. Each column of the community matrix should represent a unique species,
while each row represents a single unique survey.
}
\examples{
data(animal_observations)
data(animal_surveys)

# filter animal observations to taxon of interest
birds <- filter_obs(observations = animal_observations,
                   survey_ref = animal_surveys,
                   specify_taxon = "Aves")

# convert animal observations to community matrix
birds <- as.data.frame.matrix(xtabs(abundance ~ survey_id + species, data = birds))
birds <- cbind(survey_id = rownames(birds), birds) # convert rownames to col

# run function
set.seed(123)
exclude_simulator(birds, animal_surveys,
                 exclude_num = 15, exclude_level = "survey", # exclude 15 surveys
                 specify_cycles = 5, # exclude those in cycle 5 only
                 period_ignore = "2", # remove all data from period 2
                 rep = 10) # 10 scenarios only

}
