% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fauna.R
\docType{data}
\name{fauna}
\alias{fauna}
\title{Fauna observation data from point surveys across residential towns}
\format{
An object of class "list".
}
\source{
TBC
}
\usage{
data(fauna)
}
\description{
Data from fauna point surveys conducted in six Singapore residential
towns from 2016 to 2021. Species occurrance and abundance data were
collected for four taxons: birds, butterflies, odonates and amphibians.
}
\examples{
data(fauna)


}
\references{
TBC
}
\keyword{datasets}
