% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_obs.R
\name{filter_obs}
\alias{filter_obs}
\title{Filter species observations based on specified criteria}
\usage{
filter_obs(
  observations,
  survey_ref,
  specify_area = NULL,
  specify_period = NULL,
  specify_taxon = NULL,
  survey_id = "survey_id",
  point_id = "point_id",
  area = "town",
  period = "round",
  taxon = "priority"
)
}
\arguments{
\item{observations}{Dataframe of species observations.
It should include columns for \code{survey_id}, \code{town}, \code{round}, \code{priority}.}

\item{survey_ref}{Dataframe of all surveys conducted. Values in the column
\code{survey_id} should correspond to those in the \code{observations}.}

\item{specify_area}{Specify the \code{area}(s). Defaults to \code{NULL}, which includes all towns.}

\item{specify_period}{Specify the survey \code{period}(s). Defaults to \code{NULL}, which includes all rounds.}

\item{specify_taxon}{Specify the \code{taxon} taxon/taxa. Defaults to \code{NULL}, which includes all taxa.}

\item{survey_id}{Column name of the unique identifier for each survey in \code{observations} and
\code{survey_ref}. Defaults to \code{survey_id}.}

\item{point_id}{Column name of the unique identifier for each point in \code{observations} and
\code{survey_ref}. Defaults to \code{point_id}.}

\item{area}{Column name of the town specified in \code{observations} and \code{survey_ref}.
Defaults to \code{town}.}

\item{period}{Column name of the sampling round specified in \code{observations} and \code{survey_ref}.
Defaults to \code{round}.}

\item{taxon}{Column name of the priority taxon specified in \code{observations} and \code{survey_ref}.
Defaults to \code{priority}.}
}
\value{
The dataframe \code{observations} subset according to specified criteria.
}
\description{
Filters species observations from biodiversity surveys based on specified criteria.
}
