% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_observations.R
\name{filter_observations}
\alias{filter_observations}
\title{Filter animal observations based on specified criteria}
\usage{
filter_observations(
  observations,
  survey_ref,
  specify_area = NULL,
  specify_period = NULL,
  specify_taxon = NULL,
  survey_id = "survey_id",
  point_id = "point_id",
  area = "area",
  period = "period",
  taxon = "taxon"
)
}
\arguments{
\item{observations}{Dataframe of animal species observations.
It should include columns for \code{survey_id}, \code{area}, \code{period}, \code{taxon}.}

\item{survey_ref}{Dataframe of all surveys conducted. Values in the column
\code{survey_id} should correspond to those in the \code{observations}.}

\item{specify_area}{Specify the area(s) of interest as a character string (can be a vector of multiple areas). Defaults to \code{NULL}, which includes all areas.}

\item{specify_period}{Specify the survey period(s) of interest as a character string (can be a vector of multiple periods). Defaults to \code{NULL}, which includes all periods.}

\item{specify_taxon}{Specify the taxon (animal group) of interest as a character string (can be a vector of multiple taxa). Defaults to \code{NULL}, which includes all taxa.}

\item{survey_id}{Column name of the unique identifier for each survey in \code{observations} and
\code{survey_ref}. Defaults to \code{"survey_id"}.}

\item{point_id}{Column name of the unique identifier for each point in \code{observations} and
\code{survey_ref}. Defaults to \code{"point_id"}.}

\item{area}{Column name for the areas specified in \code{observations} and \code{survey_ref}.
Defaults to \code{"area"}.}

\item{period}{Column name for the sampling periods specified in \code{observations} and \code{survey_ref}.
Defaults to \code{"period"}.}

\item{taxon}{Column name for the taxa specified in \code{observations} and \code{survey_ref}.
Defaults to \code{"taxon"}.}
}
\value{
The dataframe \code{observations} subset according to the specified criteria.
}
\description{
Helper function to filter species observations from animal surveys based on specified criteria.
}
\examples{

data(animal_observations)
data(animal_surveys)

# filter to birds in both periods 1 and 2
filter_observations(observations = animal_observations,
                    survey_ref = animal_surveys,
                    specify_taxon = "Aves",
                    specify_period = c("1","2"))

}
