% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid.R
\name{generate_grid}
\alias{generate_grid}
\title{Generate regular grid of spatial points within target area of interest}
\usage{
generate_grid(target_areas, innerbuffer_m, pixelsize_m)
}
\arguments{
\item{target_areas}{sf object(s) of type POLYGON or MULTIPOLYGON.}

\item{innerbuffer_m}{numeric value indicating the minimum distance between generated points and the boundaries of the target area.}

\item{pixelsize_m}{numeric value for output grid cell size (in metres).}
}
\value{
An sf dataframe with an additional \code{geometry} (\code{sfc}) column of type POINT.
}
\description{
Generate a regular grid of spatial points at a specified pixel resolution (i.e. density of points),
from provided polygon(s) or multipolygon(s). The points will be used in subsequent processing,
for example, to summarise landscape variables as model inputs, as well as to generate spatial predictions
across the target area. This function is a wrapper around the \code{sf::st_make_grid()} function with the argument \code{what = "centers"}.
}
