% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genfeatures_lsm.R
\name{genfeatures_lsm}
\alias{genfeatures_lsm}
\title{Generate features based on landscape metrics at point locations}
\usage{
genfeatures_lsm(
  raster,
  points,
  buffer_sizes,
  class_names = NULL,
  class_values = NULL,
  landscape_name = NULL,
  point_id = "point_id",
  period = "period",
  what = NULL,
  level = NULL,
  ...
)
}
\arguments{
\item{raster}{SpatRaster object (\code{terra::rast()}). Classified land cover raster object to be analysed.
The number of layers should correspond to the number of survey periods present in \code{points}.}

\item{points}{Sampling points (sf object) for the calculation of landscape metrics.}

\item{buffer_sizes}{Radius of circles for sampling points (in mapunits);
landscape metrics will be calculated within the buffer area.}

\item{class_names}{Vector (character) of land cover class names to be used to identify the corresponding integer values in \code{class_values}.}

\item{class_values}{Vector of (integer) values of interest within \code{raster}. Should not include the value \code{0}.}

\item{landscape_name}{Character string for name of landscape. Used to label landscape-level metrics.}

\item{point_id}{Column name of the sampling point id within the \code{points} sf.}

\item{period}{Column name of the survey period within the \code{points} sf.
The column should contain integers that correspond to the number of layers in \code{raster}.}

\item{what}{Argument passed to \code{landscapemetrics::calculate_lsm()}.
Currently only supports class- and landscape-level metrics}

\item{level}{Argument passed to \code{landscapemetrics::calculate_lsm()}.
Currently only supports class- and landscape-level metrics.}

\item{...}{Additional arguments (e.g. \verb{type=}) passed to \code{landscapemetrics::calculate_lsm()}.}
}
\value{
A list containing the landscape metrics across the input \code{buffer_sizes}.
Each element in the list contains results for each \code{point_id} and \code{period}, for a particular buffer size.
}
\description{
Generate features by summarising landscape metrics from a classified raster,
at specific point locations and buffer radii.
}
