% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genfeatures_lsm_specified.R
\name{genfeatures_lsm_specified}
\alias{genfeatures_lsm_specified}
\title{Calculate specified landscape metrics at point locations}
\usage{
genfeatures_lsm_specified(
  raster,
  class_names,
  class_values,
  predictors_lsm,
  points,
  na_threshold = 90,
  point_id = "point_id",
  ...
)
}
\arguments{
\item{raster}{A classified SpatRaster object (\code{terra::rast()}). Pixels have integer values.}

\item{class_names}{Vector (character) of land cover class names to be used to identify the corresponding integer values in \code{class_values}.}

\item{class_values}{Vector of (integer) values of interest within the classified rasters in \code{raster}. Should not include the value \code{0}.}

\item{predictors_lsm}{Vector (character) of predictor variables to be calculated from the raster(s).
The naming format is \verb{r<radius of point buffer in metres>m_lsm_<class>_<metric>} (e.g. \code{r50m_lsm_veg_pland}).
Refer to \code{landscapemetrics::list_lsm()} for the full list of metrics.}

\item{points}{Sampling points (sf object) representing the locations to calculate the metrics.}

\item{na_threshold}{Value for calculated predictor will be \code{NA} if the percentage of raster data
within the respective point buffer areas is less than this value (\code{0} to \code{100}).}

\item{point_id}{Column name of the sampling point id within the \code{points} sf. Defaults to \code{"point_id"}.}

\item{...}{Arguments passed on to \code{genfeatures_lsm()}}
}
\value{
The \code{points} object including new columns for the variables specified in \code{predictors_osm}.
}
\description{
Summarise specified landscape metrics at point locations, for classified raster object. Calls the function \code{genfeatures_lsm()} internally.
The character vector of predictor names include the specified buffer radii within which to summarise each metric.
Refer to \code{landscapemetrics::list_lsm()} for the full list of metric names and abbreviations.
}
