% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_to_grid.R
\name{geom_to_grid}
\alias{geom_to_grid}
\title{Creates an empty grid with specified resolution for subsequent spatial visualisation of diversity patterns}
\usage{
geom_to_grid(geom, buffer_dist, grid_dim)
}
\arguments{
\item{geom}{sf geometry object, typically one or more multipolygons.}

\item{buffer_dist}{minimum distance from boundaries where grid cell centres should be.}

\item{grid_dim}{dimension in metres for each grid cell in eventual grid.}
}
\value{
The dataframe \code{bound_grid} with three columns. The third column contains S3 point objects which are the grid cell centres.
}
\description{
Creates an empty grid with specified resolution from a provided multipolygon or list of multipolygons.
}
