% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roads_osm.R
\name{get_roads_osm}
\alias{get_roads_osm}
\title{Get road lines from OpenStreetMap}
\usage{
get_roads_osm(
  place,
  date = NULL,
  dir_raw = osmextract::oe_download_directory(),
  filename = NULL,
  ...
)
}
\arguments{
\item{place}{\code{sf} object (with projected coordinate reference system). Geographical area to match with the (\code{.osm.pbf}) file in the data archive.
Argument passed to \code{osmextract::oe_match()}.}

\item{date}{Date of OSM data snapshot to download. Refer to https://download.geofabrik.de
for the specific dates available. Defaults to \code{NULL} (download the latest available data).}

\item{dir_raw}{character. Directory to download the raw unprocessed OSM data. Passed to
argument \code{download_directory} in \code{osmextract::oe_read()}.}

\item{filename}{character (optional). File path to export output data (GeoJSON format).}

\item{...}{Other arguments passed to \code{osmextract::oe_read()}.}
}
\value{
The processed road lines (\code{sf} object).
}
\description{
Download and process OpenStreetMap (OSM) road lines within a specified geographical \code{place},
from the \href{https://download.geofabrik.de}{Geofabrik database}. It is a wrapper around
functions in the package \href{https://docs.ropensci.org/osmextract/index.html}{\code{osmextract}}, and
processes the downloaded files for subsequent analyses. Refer to package \code{osmextract} for
more details and options for input arguments when downloading the data.
}
\details{
OSM lines are filtered by key-value attributes, where \verb{highway:} is
\code{motorway}, \code{trunk}, \code{primary}, \code{secondary}, \code{tertiary}, \code{unclassified} or \code{residential},
or their respective \verb{*_link} values.
}
