% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_raster.R
\name{heatmap_raster}
\alias{heatmap_raster}
\title{Creates heatmap with specified resolution from a provided geometry object and site-env data.}
\usage{
heatmap_raster(means, sds, mods, landscape, geom, buffer_dist, grid_dim, name)
}
\arguments{
\item{geom}{sf geometry object, typically one or more multipolygons.}

\item{buffer_dist}{minimum distance from boundaries where grid cell centres should be.}

\item{grid_dim}{dimension in metres for each grid cell in eventual grid.}
}
\value{
The dataframe \code{bound_grid} with three columns. The third column contains S3 point objects which are the grid cell centres.
}
\description{
Creates heatmap with specified resolution from a provided geometry object and site-env data.
}
