% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_perpoint_specified.R
\name{osm_perpoint_specified}
\alias{osm_perpoint_specified}
\title{Calculate OpenStreetMap metrics at sampling points specified by name}
\usage{
osm_perpoint_specified(
  vector_list,
  predictors_osm,
  building_ndsm = NULL,
  building_height = "height",
  building_levels = "levels",
  road_lanes = "lanes",
  points,
  point_id = "point_id",
  period = "period"
)
}
\arguments{
\item{vector_list}{List of sf objects for either buildings (polygons) or roads (lines).
The sequence (number) of list elements should correspond to the survey periods (an integer) present in \code{points}.}

\item{predictors_osm}{Vector (character) of predictor variables to be calculated from the vector file(s).
Naming format is formatted as \verb{<radius in metres>_osm_<metric>} (e.g. \code{r50m_osm_buildingFA_ratio}).}

\item{building_ndsm}{List of SpatRaster objects (\code{terra::rast()}) (optional). Each element is a continuous raster
of the normalised Digital Surface Model, used to calculate building heights.
The sequence (number) of list elements should correspond to the survey periods (an integer) present in \code{points}.
If absent (\code{NULL}), the column \code{building_height} is used instead. Defaults to \code{NULL}.}

\item{building_height}{Column name of the building height for sf objects in \code{vector_list}.
Defaults to \code{"height"}.}

\item{building_levels}{Column name of the number of building levels for sf objects in \code{vector_list}.
Defaults to \code{"levels"}.}

\item{road_lanes}{Column name of the number of}

\item{points}{Sampling points (sf object) representing the locations to calculate the metrics.}

\item{point_id}{Column name of the sampling point id within the \code{points} sf. Defaults to \code{"point_id"}.}

\item{period}{Column name of the survey period within the \code{points} sf. Defaults to \code{"period"}.
The column should contain integers that correspond to the number of elements in \code{vector_list}.}
}
\value{
A list containing the specified OpenStreetMap metrics summarised at the sampling points provided.
Each element of the list contains results for a survey \code{period} (represented as integers in \code{points}).
}
\description{
Summarise specified metrics of OpenStreetMap buildings/roads at sampling points.
Multiple (list of) vectors may be classified, each corresponding
to a survey period in the dataset \code{points}. The character vector of predictor names
include the specified buffer radii within which to summarise each metric.
Buildings and roads are to be calculated separately.
}
