% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-points.R
\docType{data}
\name{points}
\alias{points}
\title{Survey point dataset}
\format{
An object of class 'list'.
}
\source{
TBC
}
\usage{
data(points)
}
\description{
Desc
}
\examples{
data(points)


}
\references{
TBC
}
\keyword{datasets}
