% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-points_landscape.R
\docType{data}
\name{points_landscape}
\alias{points_landscape}
\title{Sampling points with landscape data summarised at multiple buffer radii}
\format{
Dataframe containing landscape predictors (columns).
}
\source{
Development of a Biodiversity Index for Residential Towns using
Biodiversity Field Surveys, 2016–2022. Ministry of National Development Research Fund
(MNDRF) Grant. Awarded to the National University of Singapore and the Singapore Housing & Development Board.
}
\usage{
data(points_landscape)
}
\description{
Examples of landscape data summarised at point locations where animal surveys
were conducted. These sampling points are within the six towns (areas) in Singapore,
during two sampling periods. The six towns were Punggol (PG), Queenstown (QT),
Tampines (TP), Jurong West (JW), Bishan (BS) and Woodlands (WL).
}
\details{
Data is in 'wide format', i.e. has duplicate rows for the unique identifier
of points (column \code{point_id}), owing to the presence of multiple buffer radii per point
(column \code{radius_m}, in metres). Not all landscape predictors are summarised at all buffer radii.
Column names for landscape predictors are prefixed with the following:
\itemize{
\item \code{man_}: manually mapped on-site
\item \code{osm_}: derived from OpenStreetMap data
\item \code{lsm_}: landscape metrics based on land cover classes
}
}
\examples{
head(points_landscape)

}
\seealso{
\itemize{
\item \link{sampling_points} where the animal surveys were conducted.
}
}
\keyword{datasets}
