% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_heatmap.R
\name{predict_heatmap}
\alias{predict_heatmap}
\title{Generate continuous raster of spatial predictions}
\usage{
predict_heatmap(models, recipe_data, points_topredict, pixelsize_m)
}
\arguments{
\item{models}{Model objects for a specific animal (taxon) group to be used for predictions
(output of \code{MuMIn::get.models()}).}

\item{recipe_data}{recipe_data recipe object (\code{recipes::recipe()}) containing information on how to
transform the new data to the appropriate format, prior to making predictions (e.g. scale and center variables).}

\item{points_topredict}{sf dataframe with \code{geometry} (\code{sfc}) column of type POINT,
as well as column names that correspond to landscape predictors within \code{models}.}

\item{pixelsize_m}{numeric value for output grid cell size (in metres). Must be the same
pixel resolution of the regular grid of spatial points \code{points_topredict}.}
}
\value{
A \code{terra::rast()} raster object with continuous values representing the
spatial predictions made across a regular grid (e.g. number of animal species).
}
\description{
Generate continuous raster of spatial predictions
}
