% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac_extractor.R
\name{sac_extractor}
\alias{sac_extractor}
\title{Extract species accumulation curves based on specified criteria}
\usage{
sac_extractor(
  observations,
  survey_ref,
  specify_area,
  specify_period,
  specify_taxon,
  survey_id = "survey_id",
  area = "area",
  period = "period",
  taxon = "taxon",
  species = "species",
  genus = "genus",
  family = "family",
  abundance = "abundance"
)
}
\arguments{
\item{observations}{Dataframe of species observations.
It should include columns for \code{survey_id}, \code{area}, \code{period}, \code{taxon}, \code{species} and \code{abundance}.}

\item{survey_ref}{Dataframe of all surveys conducted. Values in the column
\code{survey_id} should correspond to those in the \code{observations}.}

\item{specify_area}{Specify the \code{area}.}

\item{specify_period}{Specify the survey \code{period}.}

\item{specify_taxon}{Specify the \code{taxon} group of interest.}

\item{survey_id}{Column name of the unique identifier for each survey in \code{observations} and
\code{survey_ref}. Defaults to \code{survey_id}.}

\item{area}{Column name of the area specified in \code{observations} and \code{survey_ref}.
Defaults to \code{area}.}

\item{period}{Column name of the sampling period specified in \code{observations} and \code{survey_ref}.
Defaults to \code{period}.}

\item{taxon}{Column name of the taxon of interest specified in \code{observations} and \code{survey_ref}.
Defaults to \code{taxon}.}

\item{species}{Column name of the species specified in \code{observations}.
Defaults to \code{species}.}

\item{genus}{Column name of the genus specified in \code{observations}.
Defaults to \code{genus}.  Used to remove genus-level records in the \code{species} column,
if all species within the group are observed within the filtered dataset.}

\item{family}{Column name of the family specified in \code{observations}.
Defaults to \code{family}. Used to remove family-level records in the \code{species} column,
if all species within the group are observed within the filtered dataset.}

\item{abundance}{Column name of the species abundance specified in \code{observations}.
Defaults to \code{abundance}.}
}
\value{
A dataframe with columns for the \code{sites}, species \code{richness} and \code{sd} (standard deviation)
of results, as well as the information specified in \code{specify_area}, \code{specify_period} and \code{specify_taxon}.
}
\description{
Wrapper function to \code{filter_obs()} which filters species observations from biodiversity surveys, based on specified criteria.
Subsequently forms community matrix, and removes taxon group-level (genus/family)
if all species within group are observed in the area (& period), using the function \code{check_taxongrps()}.
Finally, runs the function \code{specaccum()} in package \code{vegan}, and extracts output
data. Data on both the biodiversity surveys and their species observations (and respective abundances)
must be provided.
}
