% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampling_areas.R
\docType{data}
\name{sampling_areas}
\alias{sampling_areas}
\title{Areas of interest where sampling points were generated}
\format{
Tabular \href{https://r-spatial.github.io/sf/index.html}{\code{sf}} object.
Each row has a geometry of type \code{POLYGON}.
}
\source{
Development of a Biodiversity Index for Residential Towns using
Biodiversity Field Surveys, 2016–2022. Ministry of National Development Research Fund
(MNDRF) Grant. Awarded to the National University of Singapore and the Singapore Housing & Development Board.
}
\usage{
data(sampling_areas)
}
\description{
Six towns (areas) in Singapore: Punggol (PG), Queenstown (QT),
Tampines (TP), Jurong West (JW), Bishan (BS) and Woodlands (WL).
}
\examples{
head(sampling_areas)

}
\seealso{
\itemize{
\item \link{sampling_points} where the animal surveys were conducted.
}
}
\keyword{datasets}
