% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampling_forests.R
\docType{data}
\name{sampling_forests}
\alias{sampling_forests}
\title{Sub-areas of interest where sampling points were generated}
\format{
Tabular \href{https://r-spatial.github.io/sf/index.html}{\code{sf}} object.
Each row has a geometry of type \code{POLYGON}.
For each sampling \code{period} (column), all forests patches within each area (town) were surveyed.
}
\source{
Development of a Biodiversity Index for Residential Towns using Biodiversity Field Surveys.
14 June 2016, 6.5 years, Ministry of National Development Research Fund (MNDRF) Grant.
Awarded to the National University of Singapore (host institution) and the Singapore Housing & Development Board.
}
\usage{
data(sampling_forests)
}
\description{
Forest patches within the six towns (areas) in Singapore: Punggol (PG), Queenstown (QT),
Tampines (TP), Jurong West (JW), Bishan (BS) and Woodlands (WL).
}
\examples{
head(sampling_forests)

}
\seealso{
\itemize{
\item \link{sampling_areas} are the encompassing areas of interest where these forest patches are located within.
}
}
\keyword{datasets}
