% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampling_points.R
\docType{data}
\name{sampling_points}
\alias{sampling_points}
\title{Sampling points where animal surveys were conducted}
\format{
\code{sf} points
}
\source{
TBC
}
\usage{
data(sampling_points)
}
\description{
Point locations where surveys were conducted within the six towns in Singapore:
Punggol (PG), Queenstown (QT), Tampines (TP), Jurong West (JW), Bishan (BS)
and Woodlands (WL).
}
\examples{
head(sampling_points)

}
\references{
TBC
}
\seealso{
\itemize{
\item \link{sampling_areas} where these sampling points were generated within.
\item \link{animal_surveys} containing information about the surveys conducted at these sampling points.
}
}
\keyword{datasets}
