% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-surveys.R
\docType{data}
\name{surveys}
\alias{surveys}
\title{Survey info data to accompany fauna observations}
\format{
An object of class 'list'.
}
\source{
TBC
}
\usage{
data(surveys)
}
\description{
Survey information data of fauna point surveys conducted in six Singapore
residential towns from 2016 to 2021. Time and weather information of all
surveys present in the observation data set is recorded here.
}
\examples{
data(surveys)


}
\references{
TBC
}
\keyword{datasets}
