% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_otsu.R
\name{threshold_otsu}
\alias{threshold_otsu}
\title{Calculate pixel value threshold of an image via Otsu's method}
\usage{
threshold_otsu(x = NULL, range = NULL, levels = 256, ...)
}
\arguments{
\item{x}{filename to be imported using \code{terra::rast()}, or a SpatRaster object.}

\item{range}{Numeric vector (of length 2) specifying the histogram range to be used
for thresholding. Defaults to the minimum and maximum values of the imported raster.}

\item{levels}{Number of histogram bins used to calculate the threshold value,
typically based on the bit depth of the image (e.g. 8-bit image has \code{2^8} = \code{256} levels).
Defaults to \code{256}.}

\item{...}{Optional arguments to be passed to \code{terra::rast()} when importing image.}
}
\value{
Otsu's threshold value for the image (a single number), which can subsequently be
used for image classification (e.g. convert greyscale to binary image).
}
\description{
Calculate Otsu's threshold value that separates a greyscale image into two distinct classes.
The threshold value is determined by minimizing the combined intra-class variance.
\code{library(terra)} and \code{library(stars)} are used to perform out-of-memory operations.
}
\references{
Otsu, N. (1979). A threshold selection method from gray-level histograms.
IEEE transactions on systems, man, and cybernetics, 9(1), 62-66.
Pau, G., Fuchs, F., Sklyar, O., Boutros, M., & Huber, W. (2010).

EBImage—an R package for image processing with applications to cellular phenotypes.
Bioinformatics, 26(7), 979-981.
}
