% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_newdata.R
\name{validate_newdata}
\alias{validate_newdata}
\title{Validate the accuracy of provided models using new data from animal surveys}
\usage{
validate_newdata(models, recipe_data, newdata, response_var = "sprich")
}
\arguments{
\item{models}{Model objects for a specific animal (taxon) group to be used for predictions
(output of \code{MuMIn::get.models()}).}

\item{recipe_data}{recipe object (\code{recipes::recipe()}) containing information on how to
transform the new data to the appropriate format, prior to making predictions (e.g. scale and center variables).}

\item{newdata}{New data collected at point locations for the specific taxon group.
Used to validate the accuracy of the models. Should contain the columns for landscape predictors found in the models.}

\item{response_var}{Column name of the response variable specified in \code{newdata}.
Defaults to \code{sprich} (species richness).}
}
\value{
A dataframe containing the predictions (column \code{mod.avg.pred}) and other
accuracy metrics when compared to the actual data (column name as defined by the \code{response_var} argument).
}
\description{
For a specific animal (taxon) group, predict the number of species (richness) at each
point location (rows in \code{newdata}) based on the landscape predictors (columns in \code{newdata})
summarised at these locations. Comparison with the actual data from animal surveys are used to
calculate the accuracy (error rate) of the models provided.
}
