% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animal_observations.R
\docType{data}
\name{animal_observations}
\alias{animal_observations}
\title{Animal observation records from point surveys}
\format{
Dataframe
}
\source{
TBC
}
\usage{
data(animal_observations)
}
\description{
Time-based point observations of species abundance at sampling points
within six towns (areas) in Singapore: Punggol (PG), Queenstown (QT),
Tampines (TP), Jurong West (JW), Bishan (BS) and Woodlands (WL).
}
\details{
The sampling points (\code{point_id}) were randomly generated within each town (\code{area}).
Four types of animal surveys were conducted at sampling points,
corresponding to four animal groups (column \code{taxon}): Birds (Aves), Butterflies (Lepidoptera),
Odonates (Odonata) and Amphibians (Amphibia). In addition to the name of the \code{species} and
\code{abundance} (number observed) at the particular time point, the names of the \code{family} and \code{genus}
are also made available.

Each sampling \code{period} stretched across a year-long duration, where sampling points
were sampled once every two months (\code{cycle} adds up to 6 per \code{period}).
Some points were sampled again in the second period (column \code{resampled} is \code{TRUE}).
All surveys took place between the years 2016 to 2022.
}
\examples{
head(animal_observations)

}
\references{
TBC
}
\seealso{
\itemize{
\item \link{animal_surveys} containing information about the surveys conducted.
\item \link{sampling_points} where these animal surveys were conducted.
\item \link{sampling_areas} where the sampling points were generated within.
}
}
\keyword{datasets}
