% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animal_observations.R
\docType{data}
\name{animal_observations}
\alias{animal_observations}
\title{Fauna observation data from point surveys across residential towns}
\format{
An object of class 'list'.
}
\source{
TBC
}
\usage{
data(animal_observations)
}
\description{
Data from fauna point surveys conducted in six Singapore residential
towns from 2016 to 2021. Species occurrence and abundance data were
collected for four taxa: birds, butterflies, odonates and amphibians.
}
\examples{
data(animal_observations)
head(animal_observations)

}
\references{
TBC
}
\keyword{datasets}
