% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animal_surveys.R
\docType{data}
\name{animal_surveys}
\alias{animal_surveys}
\title{Animal surveys conducted at sampling points}
\format{
Dataframe
}
\source{
TBC
}
\usage{
data(animal_surveys)
}
\description{
Details about each animal survey conducted (years 2016 to 2022) at sampling points
within six towns in Singapore: Punggol (PG), Queenstown (QT), Tampines (TP),
Jurong West (JW), Bishan (BS) and Woodlands (WL). Four types of animal surveys were conducted,
corresponding to four animal groups (taxa): Birds, butterflies, odonates and amphibians.
}
\examples{
head(animal_surveys)

}
\references{
TBC
}
\seealso{
\itemize{
\item \link{sampling_points} where these animal surveys were conducted.
\item \link{sampling_areas} where the sampling points were generated within.
}
}
\keyword{datasets}
