% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animal_surveys.R
\docType{data}
\name{animal_surveys}
\alias{animal_surveys}
\title{Animal surveys conducted at sampling points}
\format{
Dataframe
}
\source{
TBC
}
\usage{
data(animal_surveys)
}
\description{
Details about each animal survey conducted at sampling points
within six towns (areas) in Singapore: Punggol (PG), Queenstown (QT), Tampines (TP),
Jurong West (JW), Bishan (BS) and Woodlands (WL).
}
\details{
The sampling points (\code{point_id}) were randomly generated within each town (\code{area}).
Four types of animal surveys were conducted at sampling points,
corresponding to four animal groups (column \code{taxon}): Birds (Aves), Butterflies (Lepidoptera),
Odonates (Odonata) and Amphibians (Amphibia).

Each sampling \code{period} stretched across a year-long duration, where sampling points
were sampled once every two months (\code{cycle} adds up to 6 per \code{period}).
Some points were sampled again in the second period (column \code{resampled} is \code{TRUE}).
All surveys took place between the years 2016 to 2022.
}
\examples{
head(animal_surveys)

}
\references{
TBC
}
\seealso{
\itemize{
\item \link{sampling_points} where these animal surveys were conducted.
\item \link{sampling_areas} where the sampling points were generated within.
}
}
\keyword{datasets}
