% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_simulator.R
\name{exclude_simulator}
\alias{exclude_simulator}
\title{Simulate multiple scenarios with randomly excluded surveys}
\usage{
exclude_simulator(
  community,
  survey_ref,
  exclude_level = "survey",
  exclude_num,
  specify_towns = NULL,
  specify_points = NULL,
  specify_rounds = NULL,
  specify_cycles = NULL,
  town_ignore = NULL,
  round_ignore = NULL,
  cycle_ignore = NULL,
  survey_id = "survey_id",
  point_id = "point_id",
  town = "town",
  round = "round",
  cycle = "cycle",
  rep = 100
)
}
\arguments{
\item{community}{Dataframe of the species community matrix (one survey per row).
It should include columns for \code{survey_id}, \code{point_id}, \code{town}, and \code{cycle}.}

\item{survey_ref}{Dataframe of all surveys conducted. Values in the column
\code{survey_id} should correspond to those in the \code{community} data.}

\item{exclude_level}{Exclude randomly based on one of two levels: \code{"survey"} or \code{"point"}.}

\item{exclude_num}{Number of surveys to exclude. Value should be a single integer.}

\item{specify_towns}{Specify the \code{town}(s) to randomly exclude surveys from.
Defaults to \code{NULL}, which excludes surveys randomly across all towns.}

\item{specify_points}{Specify the \code{point_id}(s) of sampling points to randomly exclude surveys from.
Defaults to \code{NULL}, which excludes surveys randomly across all points.}

\item{specify_rounds}{Specify the survey \code{round} to exclude surveys from. Defaults to \code{NULL},
which excludes surveys randomly across all rounds.}

\item{specify_cycles}{Specify the survey \code{cycle}(s) to exclude surveys from. Defaults to \code{NULL},
which excludes surveys randomly across all cycles.}

\item{town_ignore}{Ignore specified \code{town}(s) in \code{community} data from analysis and results. Defaults to \code{NULL},
which includes all towns in the simulation.}

\item{round_ignore}{Ignore specified \code{round} in \code{community} data from analysis and results. Defaults to \code{NULL},
which includes all rounds in the simulation.}

\item{cycle_ignore}{Ignore specified \code{cycles}s in \code{community} data from analysis and results. Defaults to \code{NULL},
which includes all cycles in the simulation.}

\item{survey_id}{Column name of the unique identifier for each survey in \code{community} and
\code{survey_ref}. Defaults to \code{survey_id}.}

\item{point_id}{Column name of the unique identifier for each sampling point in \code{survey_ref}.
Defaults to \code{survey_id}.}

\item{town}{Column name of the town specified in \code{survey_ref}.
Defaults to \code{town}.}

\item{round}{Column name of the survey round specified in \code{survey_ref}.
Defaults to \code{round}.}

\item{cycle}{Column name of the survey cycle specified in \code{survey_ref}.
Defaults to \code{cycle}.}

\item{rep}{Specify number of repetitions (scenarios) of randomly excluded survey data.
Defaults to \code{100}.}
}
\value{
A dataframe containing the count of points with each species for the full dataset,
as well as for each repetition scenario with randomly excluded data.
}
\description{
Randomly exclude either surveys or sampling points (and their associated surveys) from a
species community matrix. Each column of the community matrix should represent a unique species,
while each row represents a single unique survey (for a particular sampling point in a town,
as well as survey round and cycle).
}
