% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_obs.R
\name{filter_obs}
\alias{filter_obs}
\title{Filter species observations based on specified criteria}
\usage{
filter_obs(
  observations,
  survey_ref,
  specify_area = NULL,
  specify_period = NULL,
  specify_taxon = NULL,
  survey_id = "survey_id",
  point_id = "point_id",
  area = "area",
  period = "period",
  taxon = "taxon"
)
}
\arguments{
\item{observations}{Dataframe of species observations.
It should include columns for \code{survey_id}, \code{area}, \code{period}, \code{taxon}.}

\item{survey_ref}{Dataframe of all surveys conducted. Values in the column
\code{survey_id} should correspond to those in the \code{observations}.}

\item{specify_area}{Specify the \code{area}(s). Defaults to \code{NULL}, which includes all areas.}

\item{specify_period}{Specify the survey \code{period}(s). Defaults to \code{NULL}, which includes all periods.}

\item{specify_taxon}{Specify the \code{taxon} taxon/taxa. Defaults to \code{NULL}, which includes all taxa.}

\item{survey_id}{Column name of the unique identifier for each survey in \code{observations} and
\code{survey_ref}. Defaults to \code{survey_id}.}

\item{point_id}{Column name of the unique identifier for each point in \code{observations} and
\code{survey_ref}. Defaults to \code{point_id}.}

\item{area}{Column name of the area specified in \code{observations} and \code{survey_ref}.
Defaults to \code{area}.}

\item{period}{Column name of the sampling period specified in \code{observations} and \code{survey_ref}.
Defaults to \code{period}.}

\item{taxon}{Column name of the taxon of interest specified in \code{observations} and \code{survey_ref}.
Defaults to \code{taxon}.}
}
\value{
The dataframe \code{observations} subset according to specified criteria.
}
\description{
Filters species observations from biodiversity surveys based on specified criteria.
}
