% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac_extractor.R
\name{sac_extractor}
\alias{sac_extractor}
\title{Extract species accumulation curves based on specified criteria}
\usage{
sac_extractor(
  observations,
  survey_ref,
  specify_town,
  specify_round,
  specify_priority,
  survey_id = "survey_id",
  town = "town",
  round = "round",
  priority = "priority",
  species = "species",
  genus = "genus",
  family = "family",
  abundance = "abundance"
)
}
\arguments{
\item{observations}{Dataframe of species observations.
It should include columns for \code{survey_id}, \code{town}, \code{round}, \code{priority}, \code{species} and \code{abundance}.}

\item{survey_ref}{Dataframe of all surveys conducted. Values in the column
\code{survey_id} should correspond to those in the \code{observations}.}

\item{specify_town}{Specify the \code{town}.}

\item{specify_round}{Specify the survey \code{round}.}

\item{specify_priority}{Specify the \code{priority} taxon.}

\item{survey_id}{Column name of the unique identifier for each survey in \code{observations} and
\code{survey_ref}. Defaults to \code{survey_id}.}

\item{town}{Column name of the town specified in \code{observations} and \code{survey_ref}.
Defaults to \code{town}.}

\item{round}{Column name of the sampling round specified in \code{observations} and \code{survey_ref}.
Defaults to \code{round}.}

\item{priority}{Column name of the priority taxon specified in \code{observations} and \code{survey_ref}.
Defaults to \code{priority}.}

\item{species}{Column name of the species specified in \code{observations}.
Defaults to \code{species}.}

\item{genus}{Column name of the genus specified in \code{observations}.
Defaults to \code{genus}.  Used to remove genus-level records in the \code{species} column,
if all species within the group are observed within the filtered dataset.}

\item{family}{Column name of the family specified in \code{observations}.
Defaults to \code{family}. Used to remove family-level records in the \code{species} column,
if all species within the group are observed within the filtered dataset.}

\item{abundance}{Column name of the species abundance specified in \code{observations}.
Defaults to \code{abundance}.}
}
\value{
A dataframe with columns for the \code{sites}, species \code{richness} and \code{sd} (standard deviation)
of results, as well as the information specified in \code{specify_town}, \code{specify_round} and \code{specify_priority}.
}
\description{
Wrapper function to \code{filter_obs()} which filters species observations from biodiversity surveys, based on specified criteria.
Subsequently forms community matrix, and removes taxon group-level (genus/family)
if all species within group are observed in the town (& round), using the function \code{check_taxongrps()}.
Finally, runs the function \code{specaccum()} in package \code{vegan}, and extracts output
data. Data on both the biodiversity surveys and their species observations (and respective abundances)
must be provided.
}
