% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampling_areas.R
\docType{data}
\name{sampling_areas}
\alias{sampling_areas}
\title{Survey area boundaries}
\format{
An object of class 'list'.
}
\source{
TBC
}
\usage{
data(sampling_areas)
}
\description{
Desc
}
\examples{
head(sampling_areas)

}
\references{
TBC
}
\keyword{datasets}
