% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampling_areas.R
\docType{data}
\name{sampling_areas}
\alias{sampling_areas}
\title{Areas of interest where sampling points were generated within}
\format{
\code{sf} polygons
}
\source{
TBC
}
\usage{
data(sampling_areas)
}
\description{
Six residential towns in Singapore: Punggol (PG), Queenstown (QT),
Tampines (TP), Jurong West (JW), Bishan (BS) and Woodlands (WL).
}
\examples{
head(sampling_areas)

}
\references{
TBC
}
\seealso{
\itemize{
\item \link{sampling_points} where the animal surveys were conducted.
}
}
\keyword{datasets}
