% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampling_points.R
\docType{data}
\name{sampling_points}
\alias{sampling_points}
\title{Survey point dataset}
\format{
An object of class 'list'.
}
\source{
TBC
}
\usage{
data(sampling_points)
}
\description{
Desc
}
\examples{
data(sampling_points)
head(sampling_points)

}
\references{
TBC
}
\keyword{datasets}
