% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampling_points.R
\docType{data}
\name{sampling_points}
\alias{sampling_points}
\title{Sampling points where animal surveys were conducted}
\format{
Tabular \href{https://r-spatial.github.io/sf/index.html}{\code{sf}} object.
Each row has a geometry of type \code{POINT}.
}
\source{
Development of a Biodiversity Index for Residential Towns using Biodiversity Field Surveys.
14 June 2016, 6.5 years, Ministry of National Development Research Fund (MNDRF) Grant, $1,286,979.79.
Awarded to the National University of Singapore (host institution) and the Singapore Housing Development Board.
}
\usage{
data(sampling_points)
}
\description{
Point locations where animal surveys were conducted within the six towns (areas) in Singapore,
during two sampling periods. The six towns were Punggol (PG), Queenstown (QT), Tampines (TP), Jurong West (JW), Bishan (BS)
and Woodlands (WL).
}
\details{
Points were randomly generated within each town (\code{area}),
stratified by \code{landcover} type (either \code{Urban} or \code{Forest}).
Each sampling \code{period} stretched across a year-long duration;
some points were sampled in both periods.
}
\examples{
head(sampling_points)

}
\seealso{
\itemize{
\item \link{sampling_areas} where these sampling points were generated within.
\item \link{animal_surveys} containing information about the surveys conducted at these sampling points.
}
}
\keyword{datasets}
